feat: split code logic & add engine to standardize code
