+++ src/generator/terminal_interaction.rs
@@ -1,0 +1,11 @@
+pub mod terminal_intercation {
+
+    pub fn print_menu() {
+        print!("+-+-+-+-+ ORION SHIELD +-+-+-+-+\n");
+        print!("Password generator\n\n");
+    }
+
+    pub fn print_config() {
+        print!("should print config");
+    }
+}
