+++ src/lib.rs
@@ -1,115 +1,4 @@
-mod terminal_intercation {
-    // used to handle user actions
-    enum Command {
-        Config,
-        Generate,
-        Quit,
-    }
-
-    pub fn print_menu() {
-        print!("ORION SHIELD");
-        print!("Password generator");
-    }
-
-    pub fn print_config() {
-        print!("should print config");
-    }
-}
-
-pub mod generator {
-    use rand::{thread_rng, Rng};
-
-    use crate::utils;
-
-    pub struct Password {
-        length: u8,
-        with_upper_case: bool,
-        with_spec_char: bool,
-        with_digit: bool,
-    }
-
-    // (lenght, got upper case, got special char, got digit
-    pub type PasswordConfig = (u8, bool, bool, bool);
-
-    impl Password {
-        pub fn new(cfg: PasswordConfig) -> Password {
-            Password {
-                length: cfg.0,
-                with_upper_case: cfg.1,
-                with_spec_char: cfg.2,
-                with_digit: cfg.3,
-            }
-        }
-
-        pub fn get_config(&self) -> PasswordConfig {
-            (
-                self.length,
-                self.with_upper_case,
-                self.with_spec_char,
-                self.with_digit,
-            )
-        }
-
-        pub fn update_config(&mut self, cfg: PasswordConfig) -> Result<String, String> {
-            // config validation
-            if let Err(conf) = config_valid(&cfg) {
-                return Err(conf);
-            }
-
-            self.length = cfg.0;
-            self.with_upper_case = cfg.1;
-            self.with_spec_char = cfg.2;
-            self.with_digit = cfg.3;
-
-            Ok(String::from("config updated"))
-        }
-
-        pub fn generate(&self) -> String {
-            let mut rng = thread_rng();
-            let mut result: String = String::from("");
-
-            for _i in 0..self.length {
-                if self.with_upper_case && rng.gen() {
-                    let n = utils::ContentCategory::UpperCase;
-                    let new_s = n.pick_random_at_index();
-                    result.push_str(&new_s);
-
-                    continue;
-                }
-
-                if self.with_digit && rng.gen() {
-                    let n = utils::ContentCategory::Digit;
-                    let new_s = n.pick_random_at_index();
-                    result.push_str(&new_s);
-
-                    continue;
-                }
-
-                if self.with_spec_char && rng.gen() {
-                    let n = utils::ContentCategory::SpecialChars;
-                    let new_s = n.pick_random_at_index();
-                    result.push_str(&new_s);
-
-                    continue;
-                }
-
-                let n = utils::ContentCategory::Lowercase;
-                let new_s = n.pick_random_at_index();
-                result.push_str(&new_s);
-            }
-
-            result
-        }
-    }
-
-    fn config_valid(cfg: &PasswordConfig) -> Result<&PasswordConfig, String> {
-        if cfg.0 < 4 {
-            return Err(String::from("password length could not be below 4"));
-        }
-
-        Ok(cfg)
-    }
-}
+pub mod generator;
 
 mod utils {
     use rand::{thread_rng, Rng};
@@ -174,12 +63,12 @@ mod utils {
 
 #[cfg(test)]
 mod tests {
-    use crate::generator;
+    use crate::generator::password::{Password, PasswordConfig};
 
     #[test]
     fn should_create_pwd() {
-        let cfg: generator::PasswordConfig = (8, false, true, false);
-        let pwd: generator::Password = generator::Password::new(cfg);
+        let cfg: PasswordConfig = (8, false, true, false);
+        let pwd: Password = Password::new(cfg);
 
         let test_cfg = pwd.get_config();
 
@@ -188,10 +77,10 @@ mod tests {
 
     #[test]
     fn should_update_pwd_config() {
-        let default_cfg: generator::PasswordConfig = (8, false, true, false);
-        let mut pwd: generator::Password = generator::Password::new(default_cfg);
+        let default_cfg: PasswordConfig = (8, false, true, false);
+        let mut pwd: Password = Password::new(default_cfg);
 
-        let new_cfg: generator::PasswordConfig = (18, true, true, true);
+        let new_cfg: PasswordConfig = (18, true, true, true);
 
         let res = pwd.update_config(new_cfg);
 
@@ -207,10 +96,10 @@ mod tests {
 
     #[test]
     fn should_not_update_pwd_config() {
-        let default_cfg: generator::PasswordConfig = (8, false, true, false);
-        let mut pwd: generator::Password = generator::Password::new(default_cfg);
+        let default_cfg: PasswordConfig = (8, false, true, false);
+        let mut pwd: Password = Password::new(default_cfg);
 
-        let new_cfg: generator::PasswordConfig = (2, true, false, true);
+        let new_cfg: PasswordConfig = (2, true, false, true);
 
         let res = pwd.update_config(new_cfg);
 
@@ -226,8 +115,8 @@ mod tests {
 
     #[test]
     fn should_generate_random_pwd() {
-        let cfg: generator::PasswordConfig = (8, false, true, false);
-        let pwd: generator::Password = generator::Password::new(cfg);
+        let cfg: PasswordConfig = (8, false, true, false);
+        let pwd: Password = Password::new(cfg);
 
         let gen_pwd1 = pwd.generate();
         let gen_pwd2 = pwd.generate();
