+++ src/main.rs
@@ -1,10 +1,48 @@
-use orion_shield::generator::{Password, PasswordConfig};
+use orion_shield::generator::{
+    engine::{Command, Engine},
+    password::PasswordConfig,
+    terminal_interaction,
+};
 
 fn main() {
-    let cfg: PasswordConfig = (32, true, true, true);
-    let pwd: Password = Password::new(cfg);
+    terminal_interaction::terminal_intercation::print_menu();
+    let mut engine = Engine::new();
+    let res = engine.exec(Command::GetConfig);
 
-    for _i in 0..10 {
-        print!("new pwd: {}\n", pwd.generate());
+    match res {
+        Ok(s) => print!("{}", s),
+        Err(e) => print!("error: {}", e),
+    }
+
+    let c: PasswordConfig = (17, false, true, false);
+    let p = engine.exec(Command::UpdateConfig(c));
+
+    match p {
+        Ok(s) => print!("{}", s),
+        Err(e) => print!("error: {}", e),
+    }
+
+    print!("\n\n");
+    let pwd = engine.exec(Command::Generate);
+
+    match pwd {
+        Ok(s) => print!("{}", s),
+        Err(e) => print!("error: {}", e),
+    }
+
+    print!("\n\n");
+    let res = engine.exec(Command::GetConfig);
+
+    match res {
+        Ok(s) => print!("{}", s),
+        Err(e) => print!("error: {}", e),
+    }
+
+    print!("\n\n");
+    let pwd = engine.exec(Command::Generate);
+
+    match pwd {
+        Ok(s) => print!("{}", s),
+        Err(e) => print!("error: {}", e),
     }
 }
